% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractData.R
\name{adtArcSinHCofactor}
\alias{adtArcSinHCofactor}
\title{ADT arc-sinh cofactors per sample}
\usage{
adtArcSinHCofactor(obj)
}
\arguments{
\item{obj}{[\code{list}] \cr
A named list of dgCMatrix objects. Exactly one of the names should be 'ADT'.}
}
\value{
[\code{data.frame}] \cr
 A data frame of arcsinh cofactors for each sample-antibody combination.
}
\description{
ADT arc-sinh cofactors per sample
}
