% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractData.R
\name{adtCounts}
\alias{adtCounts}
\title{ADT counts per sample}
\usage{
adtCounts(obj)
}
\arguments{
\item{obj}{[\code{list}] \cr
A named list of dgCMatrix objects. Exactly one of the names should be 'ADT'.}
}
\value{
[\code{data.frame}] \cr
 A data frame of the counts by sample and antibody.
}
\description{
ADT counts per sample
}
\author{
Trishanta Padayachee
}
