% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractData.R
\name{adtPctCounts}
\alias{adtPctCounts}
\title{ADT percentages of total counts per sample}
\usage{
adtPctCounts(obj)
}
\arguments{
\item{obj}{[\code{list}] \cr
A named list of dgCMatrix objects. Exactly one of the names should be 'ADT'.}
}
\value{
[\code{data.frame}] \cr
 A data frame of the percentage of total counts by sample and antibody.
}
\description{
ADT percentages of total counts per sample
}
\author{
Trishanta Padayachee
}
