% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractData.R
\name{adtPctScores}
\alias{adtPctScores}
\title{ADT percentages of total scores per sample and antibody}
\usage{
adtPctScores(obj)
}
\arguments{
\item{obj}{[\code{list}] \cr
A named list of dgCMatrix objects. Exactly one of the names should be 'ADT'.}
}
\value{
[\code{data.frame}] \cr
 The percentage of total scores by sample and antibody.
}
\description{
ADT percentages of total scores per sample and antibody
}
\author{
Trishanta Padayachee
}
