% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractData.R
\name{adtScore}
\alias{adtScore}
\title{Compute the ADT score of an antibody for a sample
The antibody scores are computed by counting the number of unique UMI Counts 
 (excluding zero) per sample that are observed in intervals of 100 and summing 
 this value over the intervals.}
\usage{
adtScore(x)
}
\arguments{
\item{x}{[\code{numeric}] \cr
A vector of ADT UMI counts for a single sample.}
}
\value{
[\code{numeric}] \cr
 ADT score for a particular antibody.
}
\description{
Compute the ADT score of an antibody for a sample
The antibody scores are computed by counting the number of unique UMI Counts 
 (excluding zero) per sample that are observed in intervals of 100 and summing 
 this value over the intervals.
}
\author{
Trishanta Padayachee
}
