% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractData.R
\name{arcSinHCofactor}
\alias{arcSinHCofactor}
\title{Arc-SinH cofactor}
\usage{
arcSinHCofactor(x)
}
\arguments{
\item{x}{[\code{numeric}] A vector of counts for a single sample}
}
\value{
[\code{numeric}] \cr
 The arcsinh cofactor.
}
\description{
Arc-SinH cofactor
}
\author{
Trishanta Padayachee
}
