% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractData.R
\name{assessMetrics}
\alias{assessMetrics}
\title{Assess gene expression assay metrics based on Cell Ranger alerts}
\usage{
assessMetrics(data)
}
\arguments{
\item{data}{[\code{data.frame}] \cr
A data frame of Cell Ranger gene expression metrics. Must contain a 
'Sample ID' column and metric columns with names corresponding to those used 
by the 10xGenomics Cell Ranger pipeline 
(see https://support.10xgenomics.com/single-cell-gene-expression/software/pipelines/latest/output/gex-metrics).}
}
\value{
[\code{data.frame}] \cr
 A data frame of sample status and metric alerts.
}
\description{
Assess the Cell Ranger gene expression metrics for various assays/samples 
to establish whether they meet expectations. Expectations are based on the 
10xGenomics Cell Ranger alerts.
}
