% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{colourUniversalDesign}
\alias{colourUniversalDesign}
\title{Extract colours from the colour universal design colour-blind friendly palette}
\usage{
colourUniversalDesign(n)
}
\arguments{
\item{n}{[\code{numeric(1)}] \cr
Number of colours to extract}
}
\value{
[\code{character(1)}] \cr
 A vector of length \code{n} of colours.
}
\description{
Extract colours from the colour universal design colour-blind friendly palette
}
\author{
Trishanta Padayachee
}
