% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primaryQCMetrics.R
\name{combineMetrics}
\alias{combineMetrics}
\title{Combine a list of primary QC metrics into a single data frame}
\usage{
combineMetrics(X, fillMissing = TRUE)
}
\arguments{
\item{X}{[\code{list}] \cr
Data frames of primary QC metrics.}

\item{fillMissing}{[\code{logical(1)}, default: \code{TRUE}] \cr
For data frames with missing metric column(s) (e.g., due to the absence of 
an assay), should a single zero observation be included in the (missing) 
metric columns, so that the sample's name still appears in figures?}
}
\value{
[\code{data.frame}] \cr
 A data.frame of primary QC metrics.
}
\description{
Combine a list of primary QC metrics into a single data frame
}
\author{
Trishanta Padayachee
}
