% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonParams.R
\name{commonParams}
\alias{commonParams}
\title{Arguments that are frequently used in the scReports package}
\arguments{
\item{sampleCRQC}{[\code{character}] \cr 
File paths to Cell Ranger outputted csv files containing sample-specific 
summary metrics. The order of the file paths in \code{sampleCRQC} should
correspond to the order of the sample names in \code{sampleNames}.}

\item{sampleIDs}{[\code{character}] \cr 
Sample IDs. The order of the IDs should correspond to the order of the file 
paths provided to the argument \code{samplePaths}. These IDs are typically 
less descriptive than those provided to the \code{sampleNames} argument.}

\item{sampleNames}{[\code{character}] \cr 
Sample names. The order of the names should correspond to the file paths 
provided to the argument \code{samplePaths} and/or \code{sampleCRQC}. 
These names are used to label the samples and need not be identical to the 
actual file names.}

\item{sampleName}{[\code{character(1)}] \cr 
Sample name.}

\item{samplePaths}{[\code{character}] \cr 
File paths to either 10xGenomics h5 files or to AnnData h5ad files. If 
\code{sampleNames} are also specified, the order of the file paths should 
correspond to the order of the sample names in \code{sampleNames}.}

\item{path}{[\code{character(1)}] \cr 
Absolute or relative file path.}

\item{metaFile}{[\code{character(1)}] \cr 
Path to a meta.tsv file containing meta data on all the samples analysed.}

\item{verbose}{[\code{logical(1)}, default: \code{FALSE}] \cr 
Should progress messages be displayed? The default value, \code{FALSE}, 
indicates that logging messages should be kept to a minimum.}

\item{makeFeaturesUnique}{[\code{logical(1)}, default: \code{TRUE}] \cr 
Whether to make feature names unique. The default value, \code{TRUE}, 
implies that if feature names are not unique, they should be made unique. 
This is done using the \code{base::make.unique} function. If \code{FALSE}, 
the original feature names are kept.}

\item{horizontalBarCharts}{[\code{logical(1)}, default: \code{FALSE}] \cr 
If \code{TRUE}, samples are plotted on the y-axis resulting in horizontal 
bars. This layout allows for figure heights to adjust as a function of the 
number of samples.}

\item{horizontalViolinPlots}{[\code{logical(1)}, default: \code{TRUE}] \cr 
If \code{TRUE}, samples are plotted on the y-axis. This layout allows for 
figure heights to be adjusted as a function of the number of samples.}

\item{geneFilteringMinCells}{[\code{numeric(1)}, default: \code{3}] \cr 
Genes with fewer than \code{geneFilteringMinCells} cells will be filtered 
out.}

\item{adtFilteringMinCells}{[\code{numeric(1)}, default: \code{0}] \cr 
Antibodies with fewer than \code{adtFilteringMinCells} cells will be 
filtered out. Since the absence of certain antibodies is also informative, 
the default value is \code{0}.}

\item{antibodyFilteringMinCells}{[\code{numeric(1)}, default: \code{0}] \cr 
Antibodies with fewer than \code{antibodyFilteringMinCells} cells will be 
filtered out. Since the absence of certain antibodies is also informative, 
the default value is \code{0}.}

\item{cellFilteringMinGenes}{[\code{numeric(1)}, default: \code{1}] \cr 
Cells with counts for fewer than \code{cellFilteringMinGenes} genes will be 
filtered out. This filtering is applied when only Gene Expression data is 
available for a sample.}

\item{cellFilteringMaxGenes}{[\code{numeric(1)}, default: \code{1}] \cr 
Cells with counts in greater than \code{cellFilteringMaxGenes} genes will 
be filtered out. This filtering is applied when only Gene Expression data 
is available for a sample.}

\item{cellFilteringMinCounts}{[\code{numeric(1)}, default: \code{10}] \cr 
Cells with fewer than \code{cellFilteringMinCounts} total counts will be 
filtered out. If the object contains both gene expression and antibody 
capture data, cells with fewer than \code{cellFilteringMinCounts} RNA or 
ADT counts will be filtered out.}

\item{cellFilteringMaxCounts}{[\code{numeric(1)}, default: \code{10}] \cr 
Cells with fewer than \code{cellFilteringMaxCounts} total counts will be 
filtered out. If the object contains both gene expression and antibody 
capture data, cells with fewer than \code{cellFilteringMinCounts} RNA or 
ADT counts will be filtered out.}

\item{cellFilteringMinCountsRNA}{[\code{numeric(1)}, default: \code{10}] \cr 
Cells with fewer than \code{cellFilteringMinCountsRNA} total gene counts 
will be filtered out.}

\item{cellFilteringMaxCountsRNA}{[\code{numeric(1)}, default: \code{10}] \cr 
Cells with greater than \code{cellFilteringMaxCountsRNA} total gene counts 
will be filtered out.}

\item{cellFilteringMinCountsADT}{[\code{numeric(1)}, default: \code{10}] \cr 
Cells with fewer than \code{cellFilteringMinCountsADT} total antibody 
counts will be filtered out.}

\item{cellFilteringMaxCountsADT}{[\code{numeric(1)}), default: \code{10}] \cr 
Cells with greater than \code{cellFilteringMaxCountsADT} total antibody 
counts will be filtered out.}

\item{cellFilteringMaxPercentMito}{[\code{numeric(1)}, default: \code{100}] \cr 
Cells with a percentage of mitochondrial gene counts greater than the 
specified value will be filtered out.}

\item{colours}{[\code{character}] \cr
A character vector of colours to be used for plotting.}

\item{palette}{[\code{character(1)}] \cr
Name of an \code{RColourBrewer} palette.}

\item{title}{[\code{character(1)}] \cr
Plot title.}

\item{xLabel}{[\code{character(1)}] \cr
X-axis label.}

\item{yLabel}{[\code{character(1)}] \cr
Y-axis label.}

\item{lLabel}{\cr
Legend name.}

\item{xAngle}{[\code{numeric(1)}, default: 0] \cr
Angle of text on the x-axis. A value between 0 and 360.}

\item{xAxisTextAngle}{[\code{numeric(1)}, default: 0] \cr
Angle of text on the x-axis. A value between 0 and 360.}

\item{textSize}{[\code{numeric(1)}] \cr
Value representing the font size.}

\item{addText}{[\code{logical(1)}, default: \code{TRUE}] \cr
Whether to add text to the plot.}

\item{makeInteractive}{[\code{logical(1)}, default: \code{TRUE}] \cr
Should an interactive plot be returned?}

\item{df}{[\code{list}] \cr
A \code{data.frame}.}

\item{obs}{[\code{character(1)}] \cr
Name of the observation to plot. Should correspond to a column in \code{df}.}

\item{obsX}{Name of the observation to plot on the X-axis. Should correspond 
to a column in \code{df}.}

\item{obsY}{Name of the observation to plot on the Y-axis. Should correspond 
to a column in \code{df}.}

\item{xObs}{Name of the observation to plot on the X-axis. Should correspond 
to a column in \code{df}.}

\item{yObs}{Name of the observation to plot on the Y-axis. Should correspond 
to a column in \code{df}.}

\item{xLimits}{[\code{numeric(2)}] \cr
X-axis limits.}

\item{yLimits}{[\code{numeric(2)}] \cr
Y-axis limits.}

\item{fillBy}{Name of a column in \code{df} that colours should be based on.}

\item{includeFigDimensions}{[\code{logical(1)}, default: \code{FALSE}] \cr
Whether to include figure dimensions in the R chunk options.}

\item{figWidth}{[\code{numeric}] \cr
A vector of figure widths with length equal to that of \code{plotList}.}

\item{figHeight}{[\code{numeric}] \cr
A vector of figure heights with length equal to that of \code{plotList}.}

\item{includeOutDimensions}{[\code{logical(1)}] \cr
Logical indicating whether to include output dimensions in the R chunk 
options.}

\item{outWidth}{[\code{numeric}] \cr
A vector of output widths with length equal to that of \code{plotList}.}

\item{outHeight}{[\code{numeric}] \cr
A vector of output heights with length equal to that of \code{plotList}.}

\item{tableList}{[\code{list}] \cr
Optionally, a named list of tables.}

\item{plotList}{[\code{list}] \cr
Optionally, a named list of plots.}

\item{plotPaths}{[\code{list}] \cr
Optionally, a named list of paths to saved figures. \code{plotPaths} can be 
specified instead of a \code{plotList}.}

\item{contentNames}{[\code{character}] \cr
A vector of names corresponding to the names in \code{textList}, 
\code{plotList}, \code{plotPaths} and \code{tableList}, if specified.}

\item{textList}{[\code{list}] \cr
Optionally, a named list of paragraphs. If an element of the list is 
\code{NULL} for a particular tab, no text will be displayed in that tab.}

\item{performChecks}{\code{logical(1)}, default: \code{TRUE}] \cr
Whether to perform routine parameter checks before proceeding.}
}
\description{
Arguments that are frequently used in the scReports package
}
