% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractData.R
\name{computeADTHeatmapData}
\alias{computeADTHeatmapData}
\title{Convert ADT metrics for heatmap data}
\usage{
computeADTHeatmapData(rawData, type)
}
\arguments{
\item{rawData}{[\code{list}] \cr
A named list of dgCMatrix objects (features by cells) of raw counts.}

\item{type}{[\code{character(1)}] \cr
Should be either "score", "gm", "counts" or "pctCounts".}
}
\value{
[\code{list}] \cr
 A list of the heatmap data and hierarchical clustering results of 
 the samples and antibodies.
}
\description{
Convert ADT metrics for heatmap data
}
\author{
Trishanta Padayachee
}
