% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{computeFigParams}
\alias{computeFigParams}
\title{Determine the number of samples included in each figure and the 
corresponding figure heights}
\usage{
computeFigParams(
  geom,
  nBreaks,
  nBreaksInStandardHeight,
  standardFigWidth = 8.5,
  standardFigHeight = 6.5,
  splitFigure = TRUE,
  performChecks = TRUE
)
}
\arguments{
\item{geom}{[\code{character(1)}] \cr
Type of plot; either "violin", "scatter", "faceted_scatter", or "heatmap".}

\item{nBreaks}{[\code{integer(1)}] \cr 
Number of samples that need to be plotted.}

\item{nBreaksInStandardHeight}{[\code{integer(1)}] \cr
Number of samples that fit in the standard figure width and height.}

\item{standardFigWidth}{[\code{numeric(1)}, default: \code{8.5}] \cr 
Standard figure width in inches.}

\item{standardFigHeight}{[\code{numeric(1)}, default: \code{8.5}] \cr
Standard figure height in inches.}

\item{splitFigure}{[\code{logical(1)}, default: \code{TRUE}] \cr
Whether to split samples across multiple figures if necessary.}

\item{performChecks}{\code{logical(1)}, default: \code{TRUE}] \cr
Whether to perform routine parameter checks before proceeding.}
}
\value{
A list containing the following values:
 geom: [\code{character(1)}] the type of figure (e.g., "violin", "scatter", etc.)
 multipleFigures: [\code{logical(1)}] logical indicating whether figures will be 
  split
 nFigures: [\code{numeric(1)}] number of figures  
 width: [\code{numeric(1)}] width of plots
 height: [\code{numeric(1)}] height of full-size plots
 heightRemainder: [\code{numeric(1)}] height of the last plot if 
  \code{remainingSamples} if \code{TRUE}
 remainder: [\code{logical(1)}] whether there is a remainder when 
  nSamples is divided by the sample breakLimit
 seqStart: [\code{numeric}] sequence of start positions 
 seqEnd: [\code{numeric}] sequence of end positions 
 nFullFigs: [\code{integer(1)}] number of full-sized plots
}
\description{
Determine the number of samples included in each figure and the 
corresponding figure heights
}
\author{
Trishanta Padayachee
}
