% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primaryQCMetrics.R
\name{computeMetrics}
\alias{computeMetrics}
\title{Compute the primary quality control metrics}
\usage{
computeMetrics(X, sampleName, verbose = FALSE, replaceNaN = TRUE)
}
\arguments{
\item{X}{[\code{list}] \cr
A named list of dgCMatrix objects (features by cells) of raw counts.}

\item{sampleName}{[\code{character(1)}] \cr 
Sample name.}

\item{verbose}{[\code{logical(1)}, default: \code{FALSE}] \cr 
Should progress messages be displayed? The default value, \code{FALSE}, 
indicates that logging messages should be kept to a minimum.}

\item{replaceNaN}{[\code{logical(1)}, default: TRUE] \cr
Should the \code{NaN} values in the percentMito be replaced with \code{0}?}
}
\value{
[\code{data.frame}] \cr
 A data frame of quality control metrics.
}
\description{
Computes the primary quality control (QC) metrics for a dgCMatrix object
 or list of dgCMatrix objects for a particular sample. The following QC 
 metrics are computed:  
 
 - the total number of counts per cell (nCounts_RNA and/or nCounts_ADT),  
 
 - the number of features expressed per cell (nFeature_RNA and/or 
 nFeature_ADT), and
 
 - the percentage of counts that can be attributed to mitochondrial genes
 (percent.mito)
}
\author{
Trishanta Padayachee
}
