% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createHeatmap.R
\name{createHeatmap}
\alias{createHeatmap}
\title{Create a heatmap}
\usage{
createHeatmap(
  df,
  obsX,
  obsY,
  fillBy,
  palette = "YlOrRd",
  xLabel = NULL,
  yLabel = NULL,
  lLabel = NULL,
  addText = TRUE,
  textSize = 2.5,
  xAngle = 90,
  limits = NULL,
  makeInteractive = TRUE
)
}
\arguments{
\item{df}{[\code{list}] \cr
A \code{data.frame}.}

\item{obsX}{Name of the observation to plot on the X-axis. Should correspond 
to a column in \code{df}.}

\item{obsY}{Name of the observation to plot on the Y-axis. Should correspond 
to a column in \code{df}.}

\item{fillBy}{Name of a column in \code{df} that colours should be based on.}

\item{palette}{[\code{character(1)}] \cr
Name of an \code{RColourBrewer} palette.}

\item{xLabel}{[\code{character(1)}] \cr
X-axis label.}

\item{yLabel}{[\code{character(1)}] \cr
Y-axis label.}

\item{lLabel}{\cr
Legend name.}

\item{addText}{[\code{logical(1)}, default: \code{TRUE}] \cr
Whether to add text to the plot.}

\item{textSize}{[\code{numeric(1)}] \cr
Value representing the font size.}

\item{xAngle}{[\code{numeric(1)}, default: 0] \cr
Angle of text on the x-axis. A value between 0 and 360.}

\item{limits}{[\code{numeric}, default: \code{NULL}] \cr
A vector of length two representing the range of values associated with 
the colour palette.}

\item{makeInteractive}{[\code{logical(1)}, default: \code{TRUE}] \cr
Should an interactive plot be returned?}
}
\value{
[\code{ggplot}] \cr 
 A ggplot object.
}
\description{
Create a heatmap
}
\author{
Trishanta Padayachee
}
