% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createQCReport.R
\name{createQCReport}
\alias{createQCReport}
\title{Create a QC report for single cell data based on user specified parameters}
\usage{
createQCReport(
  TA,
  project,
  experiment,
  sampleIDs,
  sampleNames,
  samplePaths,
  sampleCRQC = NULL,
  metaFile = NULL,
  analyst = NULL,
  autoRotate = TRUE,
  horizontalBarCharts = FALSE,
  horizontalViolinPlots = FALSE,
  isFilteredData = FALSE,
  cellRangerCount = TRUE,
  saveRObjects = FALSE,
  workingDirectory = "./wdir/",
  outputDirectory = "./wdir/",
  quietRendering = FALSE
)
}
\arguments{
\item{TA}{[\code{character(1)}] \cr
Therapeutic area.}

\item{project}{[\code{character(1)}] \cr
Project code and project name.}

\item{experiment}{[\code{character(1)}] \cr
Project code, experiment ID and experiment name.}

\item{sampleIDs}{[\code{character}] \cr 
Sample IDs. The order of the IDs should correspond to the order of the file 
paths provided to the argument \code{samplePaths}. These IDs are typically 
less descriptive than those provided to the \code{sampleNames} argument.}

\item{sampleNames}{[\code{character}] \cr 
Sample names. The order of the names should correspond to the file paths 
provided to the argument \code{samplePaths} and/or \code{sampleCRQC}. 
These names are used to label the samples and need not be identical to the 
actual file names.}

\item{samplePaths}{[\code{character}] \cr 
File paths to either 10xGenomics h5 files or to AnnData h5ad files. If 
\code{sampleNames} are also specified, the order of the file paths should 
correspond to the order of the sample names in \code{sampleNames}.}

\item{sampleCRQC}{[\code{character}] \cr 
File paths to Cell Ranger outputted csv files containing sample-specific 
summary metrics. The order of the file paths in \code{sampleCRQC} should
correspond to the order of the sample names in \code{sampleNames}.}

\item{metaFile}{[\code{character(1)}] \cr 
Path to a meta.tsv file containing meta data on all the samples analysed.}

\item{analyst}{[\code{character(1)}] \cr
Name of the individual generating the report.}

\item{autoRotate}{[\code{logical(1)}, default: \code{TRUE}] \cr
If \code{TRUE}, figures will auto rotate (flip) based on the number of 
samples analysed. Note, auto-rotation can only be turned off for less than 
30 samples.}

\item{horizontalBarCharts}{[\code{logical(1)}, default: \code{FALSE}] \cr 
If \code{TRUE}, samples are plotted on the y-axis resulting in horizontal 
bars. This layout allows for figure heights to adjust as a function of the 
number of samples.}

\item{horizontalViolinPlots}{[\code{logical(1)}, default: \code{TRUE}] \cr 
If \code{TRUE}, samples are plotted on the y-axis. This layout allows for 
figure heights to be adjusted as a function of the number of samples.}

\item{isFilteredData}{[\code{logical(1)}, default: \code{FALSE}] \cr
Whether \code{samplePaths} point to filtered h5ad files.
If \code{TRUE}, the QC metrics stored in the object are used (i.e., the metrics 
aren't recomputed). This is only possible for h5ad files with n_counts, 
n_genes and/or percent_mito columns in the obs slot.}

\item{cellRangerCount}{[\code{logical(1)}, default: \code{TRUE}] \cr
Logical indicating whether metrics_summary.csv files specified in 
\code{sampleCRQC} where obtained using the Cell Ranger count pipeline. 
\code{FALSE} indicates that the Cell Ranger multi pipeline was used.}

\item{saveRObjects}{[\code{logical(1)}, default: \code{FALSE}] \cr
Logical value indicating whether to save the R objects. 
Note, this can be extremely time consuming.}

\item{workingDirectory}{[\code{character(1)}] \cr
Working directory.}

\item{outputDirectory}{[\code{character(1)}] \cr
Output directory.}

\item{quietRendering}{[\code{logical(1)}, default: \code{FALSE}] \cr
An option to suppress printing of the pandoc command line.}
}
\description{
Create a QC report for single cell data based on user specified parameters
}
\author{
Trishanta Padayachee
}
