% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStandardScatterPlot.R
\name{createStandardScatterPlot}
\alias{createStandardScatterPlot}
\title{Create a scatter plot of two features}
\usage{
createStandardScatterPlot(
  df,
  obsX,
  obsY,
  colourBy = NULL,
  colours = NULL,
  shapeBy = NULL,
  shapes = NULL,
  xLabel = NA,
  yLabel = NA,
  lLabel = NULL,
  xLimits = NULL,
  yLimits = NULL,
  xTrans = "identity",
  yTrans = "identity",
  hlineY = NULL,
  hlineColourBy = NULL,
  hlineTypeBy = NULL,
  hlineType = NULL,
  xAngle = 0,
  pointSize = 3,
  flipCoordinates = FALSE
)
}
\arguments{
\item{df}{[\code{list}] \cr
A \code{data.frame}.}

\item{obsX}{Name of the observation to plot on the X-axis. Should correspond 
to a column in \code{df}.}

\item{obsY}{Name of the observation to plot on the Y-axis. Should correspond 
to a column in \code{df}.}

\item{colourBy}{[\code{character(1)}] \cr
Name of the observation by which to colour the points.}

\item{colours}{[\code{character}] \cr
A character vector of colours to be used for plotting.}

\item{shapeBy}{[\code{character(1)}] \cr
Name of the observation by which to shape the points.}

\item{shapes}{[\code{numeric}] \cr
A vector of integers representing the shapes that should be used when 
plotting points.}

\item{xLabel}{[\code{character(1)}] \cr
X-axis label.}

\item{yLabel}{[\code{character(1)}] \cr
Y-axis label.}

\item{lLabel}{\cr
Legend name.}

\item{xLimits}{[\code{numeric(2)}] \cr
X-axis limits.}

\item{yLimits}{[\code{numeric(2)}] \cr
Y-axis limits.}

\item{xTrans}{[\code{character(1)}] \cr
X-axis transformation.}

\item{yTrans}{[\code{character(1)}] \cr
Y-axis transformation.}

\item{hlineY}{[\code{numeric}] \cr
Y intercept for horizontal line.}

\item{hlineColourBy}{[\code{character}] \cr
A character vector of length equal to that of hlineY providing the 
categories for the hline colours.}

\item{hlineTypeBy}{[\code{character(1)}] \cr
A character vector of length equal to that of hlineY providing the 
categories for the hline types.}

\item{hlineType}{[\code{character(1)}] \cr
A character vector of line types (e.g., a combination of 'solid','dotted', 
'dashed', 'dotdash').}

\item{xAngle}{[\code{numeric(1)}, default: 0] \cr
Angle of text on the x-axis. A value between 0 and 360.}

\item{pointSize}{[\code{numeric(1)}] \cr
Size of the points.}

\item{flipCoordinates}{[\code{logical(1)}, default: \code{FALSE}] \cr
Whether to flip coordinates such that \code{obsX} appears on the y-axis and 
\code{obsY} appears on the x-axis.}
}
\value{
[\code{ggplot}] \cr
 A ggplot object.
}
\description{
\code{createScatterPlot} creates a scatter plot of two user specified
features.
}
\author{
Trishanta Padayachee
}
