% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createViolinPlot.R
\name{createViolinPlot}
\alias{createViolinPlot}
\title{Create a violin plot of a feature for a list of Seurat objects or a
list of meta data dataframes}
\usage{
createViolinPlot(
  df,
  obs,
  groupBy,
  horizontal = FALSE,
  colours = NULL,
  title = obs,
  xLabel = groupBy,
  yLabel = obs,
  xAxisTextAngle = 0,
  yLimits = NULL,
  xLimits = NULL,
  removeViolin = FALSE,
  removeJitter = FALSE,
  violinScale = "area",
  jitterShape = 16,
  jitterSize = 0.01,
  includeDefaultCoordinates = FALSE,
  includeDefaultPositionalScales = TRUE,
  performChecks = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{df}{[\code{list}] \cr
A \code{data.frame}.}

\item{obs}{[\code{character(1)}] \cr
Name of the observation to plot. Should correspond to a column in \code{df}.}

\item{groupBy}{[\code{character(1)}] \cr
Name of the variable by which to group the observations (\code{obs}). A 
violin plot will be created for each unique value of the \code{groupBy} 
variable.}

\item{horizontal}{[\code{logical(1)}, default: \code{NULL}] \cr
Whether to create horizontal violin plots. If \code{TRUE}, the \code{obs} 
variable will be plotted on the X-axis and the \code{groupBy} variable on 
the Y-axis.}

\item{colours}{[\code{character}] \cr
A character vector of colours to be used for plotting.}

\item{title}{[\code{character(1)}] \cr
Plot title.}

\item{xLabel}{[\code{character(1)}] \cr
X-axis label.}

\item{yLabel}{[\code{character(1)}] \cr
Y-axis label.}

\item{xAxisTextAngle}{[\code{numeric(1)}, default: 0] \cr
Angle of text on the x-axis. A value between 0 and 360.}

\item{yLimits}{[\code{numeric(2)}] \cr
Y-axis limits.}

\item{xLimits}{[\code{numeric(2)}] \cr
X-axis limits.}

\item{removeViolin}{[\code{logical(1)}, default: \code{FALSE}] \cr
If \code{TRUE}, the figure will only include jittered points.}

\item{removeJitter}{[\code{logical(1)}, default: \code{FALSE}] \cr
If \code{TRUE}, the figure will only contain violins.}

\item{violinScale}{[\code{character(1)}, default: "area"] \cr
Either "area", "width", or "count". Corresponds to the scale argument in 
\code{ggplot2::geom_violin} which indicates how the violins should be scaled. 
If "area", violins are scale to have the same area (before trimming of the 
tails). If "width", violins will have the same maximum width. If "count", 
areas are scaled in proportion to the number of observations.}

\item{jitterShape}{[\code{character or numeric}, default: 16] \cr
Corresponds to the shape aesthetic in \code{ggplot2::geom_point}.}

\item{jitterSize}{[\code{numeric}, default: 0.01] \cr
Corresponds to the size aesthetic in \code{ggplot2::geom_size}.}

\item{includeDefaultCoordinates}{[\code{logical(1)}, default: \code{TRUE}] \cr
If \code{TRUE}, a default Cartesian coordinate system is applied using either 
\code{coord_cartesian} or \code{coord_flip}. If \code{FALSE}, the default 
coordinate system is not applied.}

\item{includeDefaultPositionalScales}{[\code{logical(1)}, default: \code{TRUE}] \cr
If \code{TRUE}, the default positional scale is applied (i.e., 
\code{scale_x_discrete}. If \code{FALSE}, the default positional scale is 
not applied.}

\item{performChecks}{\code{logical(1)}, default: \code{TRUE}] \cr
Whether to perform routine parameter checks before proceeding.}

\item{verbose}{[\code{logical(1)}, default: \code{FALSE}] \cr 
Should progress messages be displayed? The default value, \code{FALSE}, 
indicates that logging messages should be kept to a minimum.}
}
\value{
[\code{ggplot}] \cr
 A ggplot object.
}
\description{
\code{createViolinPlot} Creates a violin plot for a specific feature for a
list of Seurat objects. Violin plots for each object appear in the same
figure.
}
\author{
Trishanta Padayachee
}
