% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{displayPlot}
\alias{displayPlot}
\title{Display plot}
\usage{
displayPlot(
  gg,
  startPos,
  endPos,
  flip = FALSE,
  zoom = FALSE,
  expansionFactor = 0,
  splitFigure = FALSE,
  performChecks = TRUE
)
}
\arguments{
\item{gg}{[\code{ggplot}] \cr
A ggplot object as returned from the function \code{createViolinPlot}.}

\item{startPos}{[\code{numeric}] \cr
A numeric vector representing the point along the x-axis at which each 
figure begins as returned from the function \code{computeFigParams}.}

\item{endPos}{[\code{numeric}] \cr
A numeric vector representing the point along the x-axis at which each 
figure ends as returned from the function \code{computeFigParams}.}

\item{flip}{[\code{logical{1}}, default: \code{FALSE}] \cr
Whether to flip the Cartesian coordinates such that the discrete \code{X} 
observation appears on the y-axis and the continuous \code{Y} observation 
appears on the x-axis.}

\item{zoom}{[\code{logical{1}}, default: \code{FALSE}] \cr
Whether to display a zoomed in version of the plot. Zooming is applied to 
the continuous \code{Y} observation only.}

\item{expansionFactor}{[\code{numeric{1}}, default: \code{0}] \cr
A numeric value between \code{0} and \code{1} representing the fraction of 
the y-axis range (i.e., before flipping the coordinates if \code{flip} is 
\code{TRUE}) to be subtracted from the lower y-axis limit and added to the 
upper y-axis limit.}

\item{splitFigure}{[\code{logical{1}}, default: \code{FALSE}] \cr
Whether \code{gg} should be displayed as a single figure or split into 
multiple figures.}

\item{performChecks}{\code{logical(1)}, default: \code{TRUE}] \cr
Whether to perform routine parameter checks before proceeding.}
}
\value{
A ggplot object.
}
\description{
\code{displayPlot} is used to extract a part of specific ggplot. It is 
assumed that the y-axis corresponds to a continuous variable and the x-axis 
to a discrete variable. The positional arguments \code{startPos} and 
\code{endPos} corresponds to the discrete x-axis variable. Zooming applies 
only to the y-axis. An option is included to expand the y-axis only 
(i.e., \code{xpansionFactor}).
}
\author{
Trishanta Padayachee
}
