% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractData.R
\name{.scatterExcludeTopN}
\alias{.scatterExcludeTopN}
\title{Exclude the top N points based on two variables}
\usage{
.scatterExcludeTopN(df, xObs, yObs, topNCutoff = 10)
}
\arguments{
\item{df}{[\code{list}] \cr
A \code{data.frame}.}

\item{xObs}{Name of the observation to plot on the X-axis. Should correspond 
to a column in \code{df}.}

\item{yObs}{Name of the observation to plot on the Y-axis. Should correspond 
to a column in \code{df}.}

\item{topNCutoff}{[\code{logical(1)}, default: \code{TRUE}] \cr
Number of top cells to exclude (a temporary solution
to remove the effect of potential outliers). The top N based on 
\code{featureX} and the top N based on \code{featureY} are removed.}
}
\description{
Exclude the top N points based on two variables
}
\author{
Trishanta Padayachee
}
