% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primaryQCMetrics.R
\name{extendMetrics}
\alias{extendMetrics}
\title{Fill in missing columns}
\usage{
extendMetrics(X, sampleName, metricNames)
}
\arguments{
\item{X}{[\code{data.frame}] \cr
A data.frame of primary QC metrics.}

\item{sampleName}{[\code{character(1)}] \cr
Name of sample associated with \code{X}.}

\item{metricNames}{[\code{character}] \cr
A vector of the expected metrics.}
}
\value{
[\code{data.frame}] \cr
 A data.frame of primary QC metrics containing the columns listed in 
 \code{metricNames}.
}
\description{
Fill in missing columns
}
\author{
Trishanta Padayachee
}
