% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primaryQCMetrics.R
\name{extractMetrics}
\alias{extractMetrics}
\title{Extract metrics from an h5ad file}
\usage{
extractMetrics(
  path,
  sampleName,
  verbose = FALSE,
  makeFeaturesUnique = TRUE,
  nFeatureRNACol = "n_genes",
  nFeatureADTCol = NULL,
  nCountRNACol = "n_counts",
  nCountADTCol = NULL,
  percentMitoCol = "percent_mito",
  nCellRNACol = "n_cells",
  nCellADTCol = NULL
)
}
\arguments{
\item{path}{[\code{character(1)}] \cr 
Absolute or relative file path.}

\item{sampleName}{[\code{character(1)}] \cr 
Sample name.}

\item{verbose}{[\code{logical(1)}, default: \code{FALSE}] \cr 
Should progress messages be displayed? The default value, \code{FALSE}, 
indicates that logging messages should be kept to a minimum.}

\item{makeFeaturesUnique}{[\code{logical(1)}, default: \code{TRUE}] \cr 
Whether to make feature names unique. The default value, \code{TRUE}, 
implies that if feature names are not unique, they should be made unique. 
This is done using the \code{base::make.unique} function. If \code{FALSE}, 
the original feature names are kept.}

\item{nFeatureRNACol}{[\code{character(1)}] \cr
Name of the column in the obs slot of the AnnData object that corresponds 
to the number of genes expressed per cell.}

\item{nFeatureADTCol}{[\code{character(1)}] \cr 
Name of the column in the obs slot of the AnnData object that corresponds 
to the number of antibodies per cell.}

\item{nCountRNACol}{[\code{character(1)}] \cr 
Name of the column in the obs slot of the AnnData object that corresponds 
to the total number of UMI counts per cell.}

\item{nCountADTCol}{[\code{character(1)}] \cr 
Name of the column in the obs slot of the AnnData object that corresponds 
to the total number of ADT counts per cell.}

\item{percentMitoCol}{[\code{character(1)}] \cr 
Name of the column in the obs slot of the AnnData object that corresponds 
to the percentage of mitochondrial UMI counts per cell.}

\item{nCellRNACol}{[\code{character(1)}] \cr 
Name of the column in the var slot of the AnnData object that corresponds 
to the number of cells expressing each gene.}

\item{nCellADTCol}{[\code{character(1)}] \cr 
Name of the column in the var slot of the AnnData object that corresponds 
to the number of cells with each antibody.}
}
\value{
[\code{list}] \cr
 A list of one or more sparse matrices of class dgCMatrix containing 
 gene expression data and/or antibody data.
}
\description{
Extract metrics from an h5ad file
}
\author{
Trishanta Padayachee
}
