% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filterSCData}
\alias{filterSCData}
\title{Filter features and/or cells}
\usage{
filterSCData(
  X,
  qcMetrics,
  sampleName,
  applyMultiModalCellFiltering = FALSE,
  cellFilteringMinGenes = 0,
  cellFilteringMaxGenes = Inf,
  cellFilteringMinCounts = 10,
  cellFilteringMaxCounts = Inf,
  cellFilteringMaxPercentMito = 100,
  geneFilteringMinCells = 3,
  antibodyFilteringMinCells = 0,
  cellFilteringMinCountsRNA = 10,
  cellFilteringMaxCountsRNA = Inf,
  cellFilteringMinCountsADT = 10,
  cellFilteringMaxCountsADT = Inf,
  verbose = FALSE
)
}
\arguments{
\item{X}{[\code{list}] \cr
A named list of dgCMatrix objects (features by cells) of raw counts 
for a single sample.}

\item{qcMetrics}{[\code{list}] \cr
A named list of primary quality control metrics.}

\item{sampleName}{[\code{character(1)}] \cr 
Sample name.}

\item{applyMultiModalCellFiltering}{[\code{logical(1)}, default: \code{FALSE}] \cr
Whether to apply multi-modal filtering. If \code{TRUE}, cells with 
nCounts_RNA and nCounts_ADT less than \code{cellFilteringMinCounts} and 
greater than \code{cellFilteringMaxCounts} are filtered out.}

\item{cellFilteringMinGenes}{[\code{numeric(1)}, default: \code{1}] \cr 
Cells with counts for fewer than \code{cellFilteringMinGenes} genes will be 
filtered out. This filtering is applied when only Gene Expression data is 
available for a sample.}

\item{cellFilteringMaxGenes}{[\code{numeric(1)}, default: \code{1}] \cr 
Cells with counts in greater than \code{cellFilteringMaxGenes} genes will 
be filtered out. This filtering is applied when only Gene Expression data 
is available for a sample.}

\item{cellFilteringMinCounts}{[\code{numeric(1)}, default: \code{10}] \cr 
Cells with fewer than \code{cellFilteringMinCounts} total counts will be 
filtered out. If the object contains both gene expression and antibody 
capture data, cells with fewer than \code{cellFilteringMinCounts} RNA or 
ADT counts will be filtered out.}

\item{cellFilteringMaxCounts}{[\code{numeric(1)}, default: \code{10}] \cr 
Cells with fewer than \code{cellFilteringMaxCounts} total counts will be 
filtered out. If the object contains both gene expression and antibody 
capture data, cells with fewer than \code{cellFilteringMinCounts} RNA or 
ADT counts will be filtered out.}

\item{cellFilteringMaxPercentMito}{[\code{numeric(1)}, default: \code{100}] \cr 
Cells with a percentage of mitochondrial gene counts greater than the 
specified value will be filtered out.}

\item{geneFilteringMinCells}{[\code{numeric(1)}, default: \code{3}] \cr 
Genes with fewer than \code{geneFilteringMinCells} cells will be filtered 
out.}

\item{antibodyFilteringMinCells}{[\code{numeric(1)}, default: \code{0}] \cr 
Antibodies with fewer than \code{antibodyFilteringMinCells} cells will be 
filtered out. Since the absence of certain antibodies is also informative, 
the default value is \code{0}.}

\item{cellFilteringMinCountsRNA}{[\code{numeric(1)}, default: \code{10}] \cr 
Cells with fewer than \code{cellFilteringMinCountsRNA} total gene counts 
will be filtered out.}

\item{cellFilteringMaxCountsRNA}{[\code{numeric(1)}, default: \code{10}] \cr 
Cells with greater than \code{cellFilteringMaxCountsRNA} total gene counts 
will be filtered out.}

\item{cellFilteringMinCountsADT}{[\code{numeric(1)}, default: \code{10}] \cr 
Cells with fewer than \code{cellFilteringMinCountsADT} total antibody 
counts will be filtered out.}

\item{cellFilteringMaxCountsADT}{[\code{numeric(1)}), default: \code{10}] \cr 
Cells with greater than \code{cellFilteringMaxCountsADT} total antibody 
counts will be filtered out.}

\item{verbose}{[\code{logical(1)}, default: \code{FALSE}] \cr 
Should progress messages be displayed? The default value, \code{FALSE}, 
indicates that logging messages should be kept to a minimum.}
}
\value{
[\code{list}] \cr
 A named list containing the filtered dgCMatrix objects, an updated 
 data frame of the raw QC metrics with filtering additional columns and a 
 filtered QC metrics data frame.
}
\description{
Filters cells based on user specified primary quality control metric cutoffs.
}
\author{
Trishanta Padayachee
}
