% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getFileExtensions}
\alias{getFileExtensions}
\title{Extract file extensions}
\usage{
getFileExtensions(filePaths)
}
\arguments{
\item{filePaths}{[\code{character}] \cr 
One or more file paths.}
}
\value{
A vector of file extensions.
}
\description{
Extract file extensions
}
\author{
Trishanta Padayachee
}
