% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getPathChild}
\alias{getPathChild}
\title{Get path to R markdown child document}
\usage{
getPathChild(child, report = NA)
}
\arguments{
\item{child}{[\code{character(1)}] \cr
Name of the R markdown child document.}

\item{report}{[\code{character(1)}] \cr
Name of the report folder in the package that contains the R markdown 
child documents. For example, 'qc' or 'integration'.}
}
\value{
[\code{character(1)}] \cr
 Path to the R markdown child document.
}
\description{
Get path to R markdown child document
}
