% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractData.R
\name{hierarchicalOrdering}
\alias{hierarchicalOrdering}
\title{Hierarchical clustering and ordering of observations
Uses the \code{stats::dist()} and \code{stats::hclust()} functions for 
 hierarchical clustering. Clustering is applied to the rows of the data \code{x}.}
\usage{
hierarchicalOrdering(x, distMethod = "euclidean", clustMethod = "ward.D")
}
\arguments{
\item{x}{[\code{matrix} or \code{data.frame}] \cr
A numeric matrix or data frame. Passed to \code{stats::dist}.}

\item{distMethod}{[\code{character(1)}] \cr
The distance measure to be used, e.g., "euclidean".}

\item{clustMethod}{[\code{character(1)}] \cr
The clustering (agglomeration) method, e.g., "ward.D", "ward.D2", or 
"single".}
}
\value{
[\code{list}] \cr
 A list of the \code{stats::hclust} output and \code{hclust} ordering of 
 observations.
}
\description{
Hierarchical clustering and ordering of observations
Uses the \code{stats::dist()} and \code{stats::hclust()} functions for 
 hierarchical clustering. Clustering is applied to the rows of the data \code{x}.
}
\author{
Trishanta Padayachee
}
