% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{prettyNumbers}
\alias{prettyNumbers}
\title{Convert to pretty numbers}
\usage{
prettyNumbers(x)
}
\arguments{
\item{x}{[\code{numeric}] \cr
A single numeric value or a numeric vector.}
}
\value{
[\code{character}] \cr
 A character vector of prettified numbers.
}
\description{
Convert to pretty numbers
}
