% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{read10xH5}
\alias{read10xH5}
\title{Read 10X h5 file}
\usage{
read10xH5(path, verbose = FALSE, makeFeaturesUnique = TRUE)
}
\arguments{
\item{path}{[\code{character(1)}] \cr 
Absolute or relative file path.}

\item{verbose}{[\code{logical(1)}, default: \code{FALSE}] \cr 
Should progress messages be displayed? The default value, \code{FALSE}, 
indicates that logging messages should be kept to a minimum.}

\item{makeFeaturesUnique}{[\code{logical(1)}, default: \code{TRUE}] \cr 
Whether to make feature names unique. The default value, \code{TRUE}, 
implies that if feature names are not unique, they should be made unique. 
This is done using the \code{base::make.unique} function. If \code{FALSE}, 
the original feature names are kept.}
}
\value{
[\code{list}] \cr
 A list of one or more dgCMatrix (sparse, column-oriented,
 numeric matrix) objects containing gene expression data and/or antibody 
 data.
}
\description{
Reads 10x Genomics h5 data files.
}
\author{
Trishanta Padayachee
}
