% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readQualityControlMetrics.R
\name{readCRQC}
\alias{readCRQC}
\title{Read in the Cell Ranger gene expression metrics}
\usage{
readCRQC(sampleNames, sampleCRQC, metrics = NULL)
}
\arguments{
\item{sampleNames}{[\code{character}] \cr 
Sample names. The order of the names should correspond to the file paths 
provided to the argument \code{samplePaths} and/or \code{sampleCRQC}. 
These names are used to label the samples and need not be identical to the 
actual file names.}

\item{sampleCRQC}{[\code{character}] \cr 
File paths to Cell Ranger outputted csv files containing sample-specific 
summary metrics. The order of the file paths in \code{sampleCRQC} should
correspond to the order of the sample names in \code{sampleNames}.}

\item{metrics}{[\code{data.frame}] \cr
A single data frame of the QC metrics for all samples in \code{sampleNames}. 
Order of rows in the data frame should correspond to the order of names in 
\code{sampleNames}.}
}
\value{
[\code{data.frame}] \cr
 A data frame containing the loaded gene expression metrics (columns) 
 for each sample (rows) in \code{sampleNames}.
}
\description{
\code{readCRQC} Reads in the Cell Ranger 3 gene expression metrics for
 a single sample or multiple samples.
}
\author{
Trishanta Padayachee
}
