% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{readData}
\alias{readData}
\title{Read single cell data contained in 10X H5 or H5AD files}
\usage{
readData(path, verbose = FALSE, makeFeaturesUnique = TRUE)
}
\arguments{
\item{path}{[\code{character(1)}] \cr 
Absolute or relative file path.}

\item{verbose}{[\code{logical(1)}, default: \code{FALSE}] \cr 
Should progress messages be displayed? The default value, \code{FALSE}, 
indicates that logging messages should be kept to a minimum.}

\item{makeFeaturesUnique}{[\code{logical(1)}, default: \code{TRUE}] \cr 
Whether to make feature names unique. The default value, \code{TRUE}, 
implies that if feature names are not unique, they should be made unique. 
This is done using the \code{base::make.unique} function. If \code{FALSE}, 
the original feature names are kept.}
}
\value{
[\code{list}] \cr
 A list of one or more dgCMatrix (sparse, column-oriented,
 numeric matrix) objects containing the gene expression data and/or antibody 
 data of a particular sample.
}
\description{
Read single cell data contained in 10X H5 or H5AD files
}
\author{
Trishanta Padayachee
}
