% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{readH5AD}
\alias{readH5AD}
\title{Read h5ad file}
\usage{
readH5AD(path, verbose = FALSE, makeFeaturesUnique = TRUE)
}
\arguments{
\item{path}{[\code{character(1)}] \cr 
Absolute or relative file path.}

\item{verbose}{[\code{logical(1)}, default: \code{FALSE}] \cr 
Should progress messages be displayed? The default value, \code{FALSE}, 
indicates that logging messages should be kept to a minimum.}

\item{makeFeaturesUnique}{[\code{logical(1)}, default: \code{TRUE}] \cr 
Whether to make feature names unique. The default value, \code{TRUE}, 
implies that if feature names are not unique, they should be made unique. 
This is done using the \code{base::make.unique} function. If \code{FALSE}, 
the original feature names are kept.}
}
\value{
[\code{list}] \cr
 A list of one or more sparse matrices of class dgCMatrix containing 
 gene expression data and/or antibody data.
}
\description{
\code{readH5AD} assumes that Gene Expression data is stored in the \code{X}
 slot of the anndata object and that if present, Antibody Capture data is 
 stored in the \code{obsm} slot of the anndata object under the name 
 \code{counts_antibody}.
}
\author{
Trishanta Padayachee
}
