% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{readSingleCellData}
\alias{readSingleCellData}
\title{Read single cell data contained in 10X H5 or H5AD files}
\usage{
readSingleCellData(
  sampleNames,
  samplePaths,
  verbose = FALSE,
  makeFeaturesUnique = TRUE
)
}
\arguments{
\item{sampleNames}{[\code{character}] \cr 
Sample names. The order of the names should correspond to the file paths 
provided to the argument \code{samplePaths} and/or \code{sampleCRQC}. 
These names are used to label the samples and need not be identical to the 
actual file names.}

\item{samplePaths}{[\code{character}] \cr 
File paths to either 10xGenomics h5 files or to AnnData h5ad files. If 
\code{sampleNames} are also specified, the order of the file paths should 
correspond to the order of the sample names in \code{sampleNames}.}

\item{verbose}{[\code{logical(1)}, default: \code{FALSE}] \cr 
Should progress messages be displayed? The default value, \code{FALSE}, 
indicates that logging messages should be kept to a minimum.}

\item{makeFeaturesUnique}{[\code{logical(1)}, default: \code{TRUE}] \cr 
Whether to make feature names unique. The default value, \code{TRUE}, 
implies that if feature names are not unique, they should be made unique. 
This is done using the \code{base::make.unique} function. If \code{FALSE}, 
the original feature names are kept.}
}
\value{
[\code{list}] \cr
 A named list of sparse matrices of class dgCMatrix containing the 
 gene expression data and/or antibody data of the samples 
 supplied to \code{samplePaths}.
}
\description{
Read single cell data contained in 10X H5 or H5AD files
}
\examples{
\dontrun{
names <- c("pbmc_1k_totalseqB_SI_h5ad", "pbmc_1k_totalseqB_SI_h5")
paths <- c(
 "./testdata/pbmc_1k_totalseqB_SI/pbmc_1k_totalseqB_SI_raw_feature_bc_matrix.h5ad",
 "./testdata/pbmc_1k_totalseqB_SI/pbmc_1k_totalseqB_SI_raw_feature_bc_matrix.h5"
)
data <- readSingleCellData(
    sampleNames = names,
    samplePaths = paths, 
    verbose = TRUE
)
}
}
\author{
Trishanta Padayachee
}
