% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabLayout.R
\name{tabLayout}
\alias{tabLayout}
\title{Create an R markdown tab layout in which each tab can contains a paragraph of 
text and/or a figure or table.}
\usage{
tabLayout(
  tabNames,
  tabLevel,
  contentNames,
  textList = NULL,
  addContentNameToTextList = TRUE,
  plotList = NULL,
  plotPaths = NULL,
  plotString = NULL,
  tableList = NULL,
  grobList = NULL,
  includeFigDimensions = FALSE,
  figWidth = NA,
  figHeight = NA,
  includeOutDimensions = FALSE,
  outWidth = NA,
  outHeight = NA
)
}
\arguments{
\item{tabNames}{[\code{character}] \cr
A vector of tab names with length equal to that of \code{contentNames}.}

\item{tabLevel}{[\code{numeric(1)}] \cr
An integer representing the level of the tab-heading.}

\item{contentNames}{[\code{character}] \cr
A vector of names corresponding to the names in \code{textList}, 
\code{plotList}, \code{plotPaths} and \code{tableList}, if specified.}

\item{textList}{[\code{list}] \cr
Optionally, a named list of paragraphs. If an element of the list is 
\code{NULL} for a particular tab, no text will be displayed in that tab.}

\item{addContentNameToTextList}{[\code{logical(1)}, default: \code{TRUE}] \cr
Whether to prepend the associated tab name to the \code{textList} entry.}

\item{plotList}{[\code{list}] \cr
Optionally, a named list of plots.}

\item{plotPaths}{[\code{list}] \cr
Optionally, a named list of paths to saved figures. \code{plotPaths} can be 
specified instead of a \code{plotList}.}

\item{plotString}{[\code{character}] \cr
Optionally, a string that contains the \code{sprintf} text for creating 
the plots in \code{plotList}.}

\item{tableList}{[\code{list}] \cr
Optionally, a named list of tables.}

\item{grobList}{[\code{list}] \cr
Optionally, a named list of grobs.}

\item{includeFigDimensions}{[\code{logical(1)}, default: \code{FALSE}] \cr
Whether to include figure dimensions in the R chunk options.}

\item{figWidth}{[\code{numeric}] \cr
A vector of figure widths with length equal to that of \code{plotList}.}

\item{figHeight}{[\code{numeric}] \cr
A vector of figure heights with length equal to that of \code{plotList}.}

\item{includeOutDimensions}{[\code{logical(1)}] \cr
Logical indicating whether to include output dimensions in the R chunk 
options.}

\item{outWidth}{[\code{numeric}] \cr
A vector of output widths with length equal to that of \code{plotList}.}

\item{outHeight}{[\code{numeric}] \cr
A vector of output heights with length equal to that of \code{plotList}.}
}
\value{
[\code{character}] \cr
 A string of code for creating the tab header and tab content.
}
\description{
Create an R markdown tab layout in which each tab can contains a paragraph of 
text and/or a figure or table.
}
\author{
Trishanta Padayachee
}
